requirements.txt
duckduckgo-search
beautifulsoup4
newspaper3k
python-dotenv
google-generativeai
requests
PyYAML